+++ src/main/java/giuseppetavella/entities/Videogioco.java
@@ -9,7 +9,7 @@ public class Videogioco extends Gioco {
     private long durataGioco;
     private GenereVideogioco genere;
     
-    public Videogioco(long idGioco, String titolo, double prezzo, long annoPubblicazione, long durataGioco,  PiattaformaVideogioco piattaformaVideogioco, GenereVideogioco genereVideogioco) {
+    public Videogioco(int idGioco, String titolo, double prezzo, long annoPubblicazione, long durataGioco,  PiattaformaVideogioco piattaformaVideogioco, GenereVideogioco genereVideogioco) {
         super(idGioco, titolo, CollezioneItemType.VIDEOGIOCO, prezzo, annoPubblicazione);
         this.setGenere(genereVideogioco);
         this.setPiattaforma(piattaformaVideogioco);
